Move user extra registry to separate module
