+++ froide/follow/forms.py
@@ -2,7 +2,7 @@ from django import forms
 from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
 
-from froide.account.forms import user_extra_registry
+from froide.account.registries import user_extra_registry
 from froide.helper.spam import SpamProtectionMixin
 
 User = get_user_model()
